refactor(video-learning-service): relocate redis and minio repository implementations
