+++ services/video_learning/design/queries/user_category_likes.sql
@@ -14,6 +14,6 @@ DELETE FROM user_category_likes WHERE user_id = $1;
 SELECT vc.name, ucl.likes
 FROM
     video_categories vc
-    LEFT JOIN user_category_likes ucl ON c.id = ucl.category_id
+    LEFT JOIN user_category_likes ucl ON vc.id = ucl.category_id
 WHERE
     ucl.user_id = $1;
