+++ services/video_learning/internal/ports/cache_repository.go
@@ -1,0 +1,14 @@
+package ports
+
+import (
+	"context"
+	"time"
+)
+
+type CacheRepository interface {
+	Set(ctx context.Context, key string, value any, expiration time.Duration) error
+	Get(ctx context.Context, key string) (string, error)
+	Delete(ctx context.Context, key string) error
+	Exists(ctx context.Context, key string) (bool, error)
+	SetNX(ctx context.Context, key string, value any, expiration time.Duration) (bool, error)
+}
